Functional and styled header and nav
